refactor: split schema runtime authority from tooling
Separate package-resolution runtime authority from schema generation and promotion tooling while keeping package-aware parsing, validation, and proof surfaces on the runtime layer.
