+++ polylogue/schemas/runtime_registry.py
@@ -1,0 +1,439 @@
+"""Runtime schema authority for package catalogs, schemas, and payload resolution."""
+
+from __future__ import annotations
+
+import copy
+import gzip
+import json
+from datetime import datetime, timezone
+from pathlib import Path
+from typing import Any
+
+from polylogue.lib.provider_identity import canonical_schema_provider as _canonical_schema_provider
+from polylogue.lib.provider_identity import normalize_provider_token
+from polylogue.paths import data_home
+from polylogue.schemas.packages import (
+    SchemaElementManifest,
+    SchemaPackageCatalog,
+    SchemaResolution,
+    SchemaVersionPackage,
+)
+from polylogue.schemas.sampling import (
+    _resolve_provider_config,
+    derive_bundle_scope,
+    extract_schema_units_from_payload,
+    profile_similarity,
+)
+from polylogue.types import Provider
+
+SCHEMA_DIR = Path(__file__).parent / "providers"
+SchemaProvider = Provider | str
+
+
+def canonical_schema_provider(provider: str | Provider) -> SchemaProvider:
+    normalized = normalize_provider_token(str(provider))
+    if not normalized:
+        return Provider.UNKNOWN
+
+    canonical = _canonical_schema_provider(normalized, default="")
+    if canonical:
+        provider_token = Provider.from_string(canonical)
+        if provider_token is not Provider.UNKNOWN:
+            return provider_token
+
+    return normalized
+
+
+class SchemaRegistry:
+    """Runtime package/catalog authority for schema resolution."""
+
+    def __init__(self, storage_root: Path | None = None):
+        self._storage_root = storage_root
+
+    @property
+    def storage_root(self) -> Path:
+        return self._storage_root if self._storage_root is not None else data_home() / "schemas"
+
+    def _provider_dir(self, provider: str | Provider) -> Path:
+        return self.storage_root / str(canonical_schema_provider(provider))
+
+    def _bundled_provider_dir(self, provider: str | Provider) -> Path:
+        return SCHEMA_DIR / str(canonical_schema_provider(provider))
+
+    def _provider_search_roots(self, provider: str | Provider) -> list[Path]:
+        roots: list[Path] = []
+        seen: set[Path] = set()
+        for candidate in (self._provider_dir(provider), self._bundled_provider_dir(provider)):
+            if candidate in seen:
+                continue
+            seen.add(candidate)
+            roots.append(candidate)
+        return roots
+
+    def _catalog_path(self, provider: str | Provider) -> Path:
+        return self._provider_dir(provider) / "catalog.json"
+
+    def _package_dir(self, provider: str | Provider, version: str) -> Path:
+        return self._provider_dir(provider) / "versions" / version
+
+    def _package_manifest_path(self, provider: str | Provider, version: str) -> Path:
+        return self._package_dir(provider, version) / "package.json"
+
+    def _catalog_path_in(self, provider_dir: Path) -> Path:
+        return provider_dir / "catalog.json"
+
+    def _provider_dir_for_catalog(self, provider: str | Provider) -> Path | None:
+        for provider_dir in self._provider_search_roots(provider):
+            if self._catalog_path_in(provider_dir).exists():
+                return provider_dir
+        return None
+
+    def _provider_dir_for_package(self, provider: str | Provider, version: str) -> Path | None:
+        for provider_dir in self._provider_search_roots(provider):
+            manifest_path = provider_dir / "versions" / version / "package.json"
+            if manifest_path.exists():
+                return provider_dir
+        return None
+
+    def load_package_catalog(self, provider: str | Provider) -> SchemaPackageCatalog | None:
+        provider_dir = self._provider_dir_for_catalog(provider)
+        if provider_dir is None:
+            return None
+        return SchemaPackageCatalog.from_dict(json.loads((provider_dir / "catalog.json").read_text(encoding="utf-8")))
+
+    def save_package_catalog(self, catalog: SchemaPackageCatalog) -> Path:
+        provider_dir = self._provider_dir(catalog.provider)
+        provider_dir.mkdir(parents=True, exist_ok=True)
+        path = self._catalog_path(catalog.provider)
+        path.write_text(json.dumps(catalog.to_dict(), indent=2), encoding="utf-8")
+        return path
+
+    def _resolve_catalog_version(self, catalog: SchemaPackageCatalog, version: str) -> str | None:
+        if version == "default":
+            return catalog.default_version or catalog.latest_version or catalog.recommended_version
+        if version == "latest":
+            return catalog.latest_version or catalog.default_version or catalog.recommended_version
+        if version == "recommended":
+            return catalog.recommended_version or catalog.default_version or catalog.latest_version
+        return version
+
+    def write_package(
+        self,
+        package: SchemaVersionPackage,
+        *,
+        element_schemas: dict[str, dict[str, Any]],
+    ) -> Path:
+        provider_token = canonical_schema_provider(package.provider)
+        package_dir = self._package_dir(provider_token, package.version)
+        elements_dir = package_dir / "elements"
+        elements_dir.mkdir(parents=True, exist_ok=True)
+
+        for element in package.elements:
+            if element.schema_file is None:
+                continue
+            schema = copy.deepcopy(element_schemas[element.element_kind])
+            schema["$id"] = f"polylogue://schemas/{provider_token}/{package.version}/{element.element_kind}"
+            schema["x-polylogue-version"] = int(package.version[1:]) if package.version.startswith("v") else package.version
+            schema["x-polylogue-package-version"] = package.version
+            schema["x-polylogue-element-kind"] = element.element_kind
+            schema["x-polylogue-registered-at"] = datetime.now(tz=timezone.utc).isoformat()
+            schema_path = elements_dir / element.schema_file
+            schema_path.write_bytes(gzip.compress(json.dumps(schema, indent=2).encode("utf-8")))
+
+        manifest_path = self._package_manifest_path(provider_token, package.version)
+        manifest_path.write_text(json.dumps(package.to_dict(), indent=2), encoding="utf-8")
+        return manifest_path
+
+    def replace_provider_packages(
+        self,
+        provider: str | Provider,
+        catalog: SchemaPackageCatalog,
+        package_schemas: dict[str, dict[str, dict[str, Any]]],
+    ) -> None:
+        provider_token = canonical_schema_provider(provider)
+        provider_dir = self._provider_dir(provider_token)
+        provider_dir.mkdir(parents=True, exist_ok=True)
+        versions_dir = provider_dir / "versions"
+        if versions_dir.exists():
+            for path in versions_dir.rglob("*"):
+                if path.is_file():
+                    path.unlink()
+            for path in sorted(versions_dir.rglob("*"), reverse=True):
+                if path.is_dir():
+                    path.rmdir()
+        versions_dir.mkdir(parents=True, exist_ok=True)
+
+        for package in catalog.packages:
+            self.write_package(package, element_schemas=package_schemas[package.version])
+        self.save_package_catalog(catalog)
+
+    def get_package(self, provider: str | Provider, version: str = "default") -> SchemaVersionPackage | None:
+        provider_token = canonical_schema_provider(provider)
+        catalog = self.load_package_catalog(provider_token)
+        if catalog is None:
+            return None
+        resolved_version = self._resolve_catalog_version(catalog, version)
+        return catalog.package(resolved_version) if resolved_version is not None else None
+
+    def get_element_schema(
+        self,
+        provider: str | Provider,
+        *,
+        version: str = "default",
+        element_kind: str | None = None,
+    ) -> dict[str, Any] | None:
+        provider_token = canonical_schema_provider(provider)
+        package = self.get_package(provider_token, version=version)
+        if package is None:
+            return None
+        element = package.element(element_kind)
+        if element is None or element.schema_file is None:
+            return None
+        provider_dir = self._provider_dir_for_package(provider_token, package.version)
+        if provider_dir is None:
+            return None
+        path = provider_dir / "versions" / package.version / "elements" / element.schema_file
+        if not path.exists():
+            return None
+        return json.loads(gzip.decompress(path.read_bytes()).decode("utf-8"))
+
+    def get_schema(self, provider: str | Provider, version: str = "default") -> dict[str, Any] | None:
+        return self.get_element_schema(provider, version=version)
+
+    def list_versions(self, provider: str | Provider) -> list[str]:
+        provider_token = canonical_schema_provider(provider)
+        catalog = self.load_package_catalog(provider_token)
+        if catalog is None:
+            return []
+        return sorted((package.version for package in catalog.packages), key=lambda value: int(value[1:]))
+
+    def list_providers(self) -> list[str]:
+        providers: set[str] = set()
+        scanned_roots: set[Path] = set()
+        for root in (self.storage_root, SCHEMA_DIR):
+            if root in scanned_roots or not root.exists():
+                continue
+            scanned_roots.add(root)
+            for path in root.iterdir():
+                if path.is_dir() and (path / "catalog.json").exists():
+                    providers.add(path.name)
+        return sorted(providers)
+
+    def _single_element_package(
+        self,
+        provider: str | Provider,
+        *,
+        version: str,
+        schema: dict[str, Any],
+        element_kind: str = "conversation_document",
+        first_seen: str | None = None,
+        last_seen: str | None = None,
+    ) -> tuple[SchemaVersionPackage, dict[str, dict[str, Any]]]:
+        provider_token = canonical_schema_provider(provider)
+        now = datetime.now(tz=timezone.utc).isoformat()
+        observed_at = schema.get("x-polylogue-generated-at")
+        profile_family_ids = [
+            str(item)
+            for item in schema.get(
+                "x-polylogue-package-profile-family-ids",
+                schema.get("x-polylogue-profile-family-ids", []),
+            )
+            if isinstance(item, str)
+        ]
+        if not profile_family_ids:
+            profile_family_ids = [
+                str(item)
+                for item in schema.get("x-polylogue-profile-family-ids", [])
+                if isinstance(item, str)
+            ]
+        element_profile_family_ids = [
+            str(item)
+            for item in schema.get("x-polylogue-profile-family-ids", profile_family_ids)
+            if isinstance(item, str)
+        ]
+        anchor_profile_family_id = str(schema.get("x-polylogue-anchor-profile-family-id", "")).strip() or next(
+            (item for item in profile_family_ids if item),
+            "",
+        )
+        observed_artifact_count = int(schema.get("x-polylogue-observed-artifact-count", 0) or 0)
+        package = SchemaVersionPackage(
+            provider=provider_token,
+            version=version,
+            anchor_kind=element_kind,
+            default_element_kind=element_kind,
+            first_seen=first_seen or observed_at or now,
+            last_seen=last_seen or first_seen or observed_at or now,
+            bundle_scope_count=0,
+            sample_count=int(schema.get("x-polylogue-sample-count", 0) or 0),
+            anchor_profile_family_id=anchor_profile_family_id,
+            profile_family_ids=profile_family_ids,
+            elements=[
+                SchemaElementManifest(
+                    element_kind=element_kind,
+                    schema_file=f"{element_kind}.schema.json.gz",
+                    sample_count=int(schema.get("x-polylogue-sample-count", 0) or 0),
+                    artifact_count=observed_artifact_count,
+                    first_seen=str(schema.get("x-polylogue-element-first-seen", observed_at or "")),
+                    last_seen=str(schema.get("x-polylogue-element-last-seen", observed_at or "")),
+                    bundle_scope_count=int(schema.get("x-polylogue-element-bundle-scope-count", 0) or 0),
+                    exact_structure_ids=[str(item) for item in schema.get("x-polylogue-exact-structure-ids", [])],
+                    profile_family_ids=element_profile_family_ids,
+                    profile_tokens=[
+                        str(item)
+                        for item in schema.get("x-polylogue-profile-tokens", [])
+                        if isinstance(item, str)
+                    ],
+                    representative_paths=[
+                        str(item)
+                        for item in schema.get("x-polylogue-representative-paths", [])
+                        if isinstance(item, str)
+                    ],
+                    observed_artifact_count=observed_artifact_count,
+                )
+            ],
+        )
+        return package, {element_kind: schema}
+
+    def register_schema(self, provider: str | Provider, schema: dict[str, Any]) -> str:
+        provider_token = canonical_schema_provider(provider)
+        versions = self.list_versions(provider_token)
+        new_version = f"v{int(versions[-1][1:]) + 1}" if versions else "v1"
+        self.write_schema_version(provider_token, new_version, schema)
+        return new_version
+
+    def write_schema_version(self, provider: str | Provider, version: str, schema: dict[str, Any]) -> Path:
+        provider_token = canonical_schema_provider(provider)
+        package, schemas = self._single_element_package(provider_token, version=version, schema=copy.deepcopy(schema))
+        catalog = self.load_package_catalog(provider_token) or SchemaPackageCatalog(provider=provider_token)
+        existing_packages = [item for item in catalog.packages if item.version != version]
+        existing_packages.append(package)
+        existing_packages.sort(key=lambda item: int(item.version[1:]))
+        catalog.packages = existing_packages
+        catalog.latest_version = existing_packages[-1].version if existing_packages else version
+        catalog.default_version = catalog.latest_version
+        catalog.recommended_version = catalog.latest_version
+        self.write_package(package, element_schemas=schemas)
+        self.save_package_catalog(catalog)
+        return self._package_dir(provider_token, version) / "elements" / f"{package.default_element_kind}.schema.json.gz"
+
+    def get_schema_age_days(self, provider: str | Provider) -> int | None:
+        package = self.get_package(provider, version="latest")
+        if package is None:
+            return None
+        try:
+            delta = datetime.now(tz=timezone.utc) - datetime.fromisoformat(package.last_seen)
+        except (ValueError, TypeError):
+            return None
+        return delta.days
+
+    def resolve_payload(
+        self,
+        provider: str | Provider,
+        payload: Any,
+        *,
+        source_path: str | None = None,
+    ) -> SchemaResolution | None:
+        provider_token = canonical_schema_provider(provider)
+        catalog = self.load_package_catalog(provider_token)
+        if catalog is None or not catalog.packages:
+            return None
+
+        config = _resolve_provider_config(provider_token)
+        units = extract_schema_units_from_payload(
+            payload,
+            provider_name=provider_token,
+            source_path=source_path,
+            raw_id=None,
+            observed_at=None,
+            config=config,
+            max_samples=64,
+        )
+        if not units:
+            default_package = self.get_package(provider_token, version="default")
+            if default_package is None:
+                return None
+            return SchemaResolution(
+                provider=str(provider_token),
+                package_version=default_package.version,
+                element_kind=default_package.default_element_kind,
+                exact_structure_id=None,
+                bundle_scope=derive_bundle_scope(provider_token, source_path),
+                reason="package_default",
+            )
+
+        unit = units[0]
+        bundle_scope = unit.bundle_scope or derive_bundle_scope(provider_token, source_path)
+
+        if bundle_scope is not None:
+            for package in catalog.packages:
+                if bundle_scope not in package.bundle_scopes:
+                    continue
+                element = package.element(unit.artifact_kind)
+                if element is None:
+                    continue
+                return SchemaResolution(
+                    provider=str(provider_token),
+                    package_version=package.version,
+                    element_kind=element.element_kind,
+                    exact_structure_id=unit.exact_structure_id,
+                    bundle_scope=bundle_scope,
+                    reason="bundle_scope",
+                )
+
+        for package in catalog.packages:
+            element = package.element(unit.artifact_kind)
+            if element is None:
+                continue
+            if unit.exact_structure_id in element.exact_structure_ids:
+                return SchemaResolution(
+                    provider=str(provider_token),
+                    package_version=package.version,
+                    element_kind=element.element_kind,
+                    exact_structure_id=unit.exact_structure_id,
+                    bundle_scope=bundle_scope,
+                    reason="exact_structure",
+                )
+
+        best_match: tuple[float, SchemaVersionPackage, SchemaElementManifest] | None = None
+        for package in catalog.packages:
+            element = package.element(unit.artifact_kind)
+            if element is None or not element.profile_tokens:
+                continue
+            score = profile_similarity(set(element.profile_tokens), set(unit.profile_tokens))
+            if best_match is None or score > best_match[0]:
+                best_match = (score, package, element)
+        if best_match is not None and best_match[0] > 0.0:
+            _score, package, element = best_match
+            return SchemaResolution(
+                provider=str(provider_token),
+                package_version=package.version,
+                element_kind=element.element_kind,
+                exact_structure_id=unit.exact_structure_id,
+                bundle_scope=bundle_scope,
+                reason="profile_family",
+            )
+
+        default_package = self.get_package(provider_token, version="default")
+        if default_package is None:
+            return None
+        return SchemaResolution(
+            provider=str(provider_token),
+            package_version=default_package.version,
+            element_kind=default_package.default_element_kind,
+            exact_structure_id=unit.exact_structure_id,
+            bundle_scope=bundle_scope,
+            reason="package_default",
+        )
+
+    def match_payload_version(
+        self,
+        provider: str | Provider,
+        payload: Any,
+        *,
+        source_path: str | None = None,
+    ) -> str | None:
+        resolution = self.resolve_payload(provider, payload, source_path=source_path)
+        return resolution.package_version if resolution is not None else None
+
+
+__all__ = ["SCHEMA_DIR", "SchemaProvider", "SchemaRegistry", "canonical_schema_provider"]
