+++ polylogue/schemas/schema_generation.py
@@ -59,13 +59,12 @@ from polylogue.schemas.semantic_inference import (
     infer_semantic_roles,
     select_best_roles,
 )
+from polylogue.schemas.shape_fingerprint import _structure_fingerprint
 from polylogue.types import Provider
 
 _HIGH_CARDINALITY_KEY_THRESHOLD = 128
 _PATHLIKE_KEY_RATIO_THRESHOLD = 0.35
 _STRUCTURE_EXEMPLARS_PER_FINGERPRINT = 8
-_FINGERPRINT_MAX_DEPTH = 8
-_FINGERPRINT_ARRAY_SAMPLE = 8
 _PROFILE_CORE_MIN_RATIO = 0.5
 _PROFILE_MAX_TOKENS = 128
 _PROFILE_SIMILARITY_THRESHOLDS = {
@@ -197,55 +196,6 @@ def _merge_schemas(schemas: list[dict[str, Any]]) -> dict[str, Any]:
     for schema in schemas:
         builder.add_schema(schema)
     return dict(builder.to_schema())
-
-
-def _structure_fingerprint(
-    value: Any,
-    *,
-    depth: int = 0,
-    max_depth: int = _FINGERPRINT_MAX_DEPTH,
-) -> Any:
-    """Build a hashable structural fingerprint for schema-dedup heuristics.
-
-    The fingerprint intentionally ignores concrete scalar values and keeps only
-    structural/type shape so repeated record variants can be collapsed before
-    feeding Genson.
-    """
-    if depth >= max_depth:
-        return ("depth-limit", type(value).__name__)
-
-    if value is None:
-        return ("null",)
-    if isinstance(value, bool):
-        return ("bool",)
-    if isinstance(value, (int, float)) and not isinstance(value, bool):
-        return ("number",)
-    if isinstance(value, str):
-        return ("string",)
-
-    if isinstance(value, list):
-        item_shapes = {
-            _structure_fingerprint(item, depth=depth + 1, max_depth=max_depth)
-            for item in value[:_FINGERPRINT_ARRAY_SAMPLE]
-        }
-        return ("array", tuple(sorted(item_shapes, key=repr)))
-
-    if isinstance(value, dict):
-        props: list[tuple[str, Any]] = []
-        for key in sorted(value):
-            child = value[key]
-            normalized_key = "*" if is_dynamic_key(key) else key
-            props.append(
-                (
-                    normalized_key,
-                    _structure_fingerprint(child, depth=depth + 1, max_depth=max_depth),
-                )
-            )
-        return ("object", tuple(props))
-
-    return ("other", type(value).__name__)
-
-
 # =============================================================================
 # Field Statistics & Schema Annotations
 # =============================================================================
@@ -1583,12 +1533,8 @@ def generate_all_schemas(
 
         if result.success and bundle.manifest is not None and bundle.catalog is not None:
             registry = SchemaRegistry(storage_root=output_dir)
-            registry.replace_provider_packages(
-                provider,
-                bundle.catalog,
-                bundle.package_schemas,
-                manifest=bundle.manifest,
-            )
+            registry.replace_provider_packages(provider, bundle.catalog, bundle.package_schemas)
+            registry.save_cluster_manifest(bundle.manifest)
 
             for legacy_name in (f"{provider}.schema.json.gz", f"{provider}.schema.json"):
                 legacy_path = output_dir / legacy_name
